pyproject.toml
[tool.poetry]
name = "customizable-continuous-integration"
version = "1.0.0"
description = "The CI automation project"
authors = ["Ryan,Gao <ryangao-au@outlook.com>"]
license = "MIT"
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#####################
#    Tool Configs   #
#####################

# black
[tool.black]
line-length = 150
include = '\.pyi?$'
force-exclude = '^(.*/|)\.?snap\_.*\.py$'


# flake8
# >>> Flake8 does not support `pyproject.toml`, use `.flake8` instead <<<

# flake8-docstring
[tool.pydocstyle]
convention = "google"

# isort
[tool.isort]
profile = "black"
skip_glob = ['**/tests/**/snap_test*.py']
line_length = 150

#############################
#    Package Requirements   #
#############################
[tool.poetry.dependencies]
python = "~=3.10.0"
dbt-core = "~1.8.5"
dbt-bigquery = "~1.8.2"
gitpython = "~3.1.43"
retrying = "~1.3.4"
numpy = "<2.0.0"

[tool.poetry.group.lint.dependencies]
pre-commit = "~3.8.0"
black = "~24.8.0"
flake8 = "~7.1.1"
flake8-docstrings = "~1.7.0"
flake8-bugbear = "~24.8.19"
flake8-builtins = "~2.5.0"
flake8-comprehensions = "~3.15.0"
pep8-naming = "~0.13.3"
isort = "~5.13.2"
mypy = "~1.11.1"
sqlfluff = "~3.1.1"
sqlfluff-templater-dbt = "~3.1.1"
shandy-sqlfmt = {extras = ["jinjafmt"], version = "~0.23.2"}

[tool.sqlfmt]
line_length = 120
dialect="polyglot"
exclude=["dbt_packages/**/*", "target/**/*.sql"]
